% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{graph}
\alias{graph}
\alias{graph_}
\title{An R6 representation of a graph data structure.}
\usage{
graph_

graph(...)
}
\arguments{
\item{...}{Arguments to pass to graph initialization.}
}
\description{
An R6 representation of a graph data structure.
}
\keyword{datasets}
