% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{graphNode}
\alias{graphNode}
\alias{graphNode_}
\title{An R6 representation of a node for a graph data structure.}
\usage{
graphNode_

graphNode(...)
}
\arguments{
\item{...}{Arguments to pass to graphNode initialization.}
}
\description{
An R6 representation of a node for a graph data structure.
}
\keyword{datasets}
