% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{graphrunner}
\alias{graphrunner}
\alias{graphrunner-package}
\title{Graphrunner is a directed graph task execution engine}
\description{
Most collections of tasks can be represented as cyclic or
acyclic directed graphs. This package makes the management
and execution of such graphs of tasks straightforward.
}
\references{
Full documentation and demos: \url{http://robertzk.github.io/graphrunner/};
  FAQ's: \url{http://robertzk.github.io/graphrunner/faq/}
}
\seealso{
The core function in this package: \code{\link{stagerunner}}. It
  defines the constructor creating graphrunner objects that allow you to
  wrap complicated task execution structures in a single object.
}
\author{
Robert Krzyzanowski <\url{http://syberia.io}>
}
