% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traversal.R
\docType{class}
\name{graphTraversalStrategy}
\alias{graphTraversalStrategy}
\alias{graphBFSTraversalStrategy}
\title{An R6 representation of a graph traversal strategy.}
\usage{
graphTraversalStrategy

graphBFSTraversalStrategy
}
\description{
A graph traversal strategy specifies how to execute tasks on
a given graph.

This graph traversal strategy begins from the boot node and
traverses "breadth-first" outward.
}
\keyword{datasets}
