% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is.graph}
\alias{is.graph}
\title{Check whether an R object is a graph object}
\usage{
is.graph(obj)
}
\arguments{
\item{obj}{any object.}
}
\value{
\code{TRUE} if the object is of class
   \code{graph}, \code{FALSE} otherwise.
}
\description{
Check whether an R object is a graph object
}
