% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is.graphNode}
\alias{is.graphNode}
\title{Check whether an R object is a graphNode object}
\usage{
is.graphNode(obj)
}
\arguments{
\item{obj}{any object.}
}
\value{
\code{TRUE} if the object is of class
   \code{graphNode}, \code{FALSE} otherwise.
}
\description{
Check whether an R object is a graphNode object
}
