% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{graph}
\alias{graph}
\alias{graph_}
\title{An R6 representation of a graph data structure.}
\usage{
graph_

graph(...)
}
\arguments{
\item{...}{Arguments to pass to graph initialization.}
}
\description{
All graphs are assumed to be connected (if they are not, use a list
of graphs instead). Any distinguished canonical node on the graph,
called the boot node, will be used to represent the entire graph
and will be the first node for purposes of traversal.
}
\keyword{datasets}
